+++ src/main/java/org/toilelibre/libe/soundtransform/model/library/note/FormulaNote.java
@@ -39,7 +39,7 @@ public abstract class FormulaNote implements Note {
     private long [] generateLongArray (final float frequency, final int nbSamples, final float sampleRate, final int maxVal, final float startAmplitude, final float endAmplitude) {
         final long [] signal = new long [nbSamples];
         for (int j = 0 ; j < nbSamples ; j++) {
-            final float coeff = (float) (((j * endAmplitude) + ((nbSamples - j) * startAmplitude)) / (1.0 * nbSamples));
+            final float coeff = (float) ((j * endAmplitude + (nbSamples - j) * startAmplitude) / (1.0 * nbSamples));
             signal [j] = (long) (this.applyFormula (j, frequency, sampleRate) * maxVal * coeff);
         }
 
@@ -48,7 +48,7 @@ public abstract class FormulaNote implements Note {
 
     private Sound generatePureNote (final float frequency, final float lengthInSeconds, final int channelnum, final float startAmplitude, final float endAmplitude) {
         final int nbSamples = (int) (this.formatInfo.getSampleRate () * lengthInSeconds * 1.0);
-        return new Sound (this.generateLongArray (frequency, nbSamples, (float)this.formatInfo.getSampleRate (), this.maxVal, startAmplitude, endAmplitude), this.formatInfo, channelnum);
+        return new Sound (this.generateLongArray (frequency, nbSamples, this.formatInfo.getSampleRate (), this.maxVal, startAmplitude, endAmplitude), this.formatInfo, channelnum);
     }
 
     @Override
